Add endpoints to API
This commit adds the endpoints to create and list dns entries.

In both cases, we have a simple delegation to previous classes present
at the code base.

There are a missing spec for list fail, but I can't see how this code
can be explored for a error since every thing at the request is a string
and we don't have required fields for listing.